web: implement add-account REST call
This call adds new account. Account is different structure from the
player and is used solely to authenticate user's credentials and determine
the privileges.

Regular accounts are associated with the player and will be used
for players to browse the ladder (and later) enter their own matches.

Administrator account is not associated with a player and will
have privileges to add matches for any players and generally
manipulate the ladder.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>