web: add support for bootstrap token
Bootstrap token is a special password for fake
bootstrap account that only works when there are no admins
in the account database. The password is a bootstrap token
value which is stored in the configuration file.

If there is a single admin account present, the bootstrap
token no longer works.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>