DESCRIPTION.rst: adding Fedora 24 installation instr./text edits
Adding installation instructions that are specific for Fedora 24;
the number of listed packages is not exhaustive - more packages
may have to be installed for a given Fedora 24 installation. In
addition, for the installation itself, it is important to use a
bash shell and use the source command (see remarks in the edited
text).

Signed-off-by: Adriaan van Wijngaarden <ajdlvw@gmail.com>