web: add new test-only REST call
The new REST call is used for testing only (for now) and determines
the player ladder for a given date. The client submits the player
ID and the date of interest and the call returns the player
ladder for that date. If the date is omitted, the current
date is used.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>