db: use player ladder on match date to determine the ladder
This patch actually makes the match entries robust to
race reporting. The ladder on match date is used to
determine the match ladder and whether the player
should be promoted or not. The same check is used
in deferred match-crediting, which also prepares the
code for having pending match entries (prerequisite
for allowing the players to enter their own matches).

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>