db: add hacky database fixer
There is a quirk in sqlite3 library. For fields that are
BOOL NOT NULL DEFAULT FALSE, default value is not the same
'false' that we get when we actually set the field to
False. In the former case, we get the string 'FALSE', in the
latter case, we get the actual (Python) boolean False.

That makes the behavior of the database inconsistet. To fix
it, run through the database at the startup time and fix all
potentially inconsistent fields.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>