web: add db parameter to the kick_season call
db parameter can have two values: set and clear. The 'set'
value sets the points that the player should have at the
start of the season. The 'clear' value resets points to zero.

For now, the intent is only to call db=set call once at the start
of the season, but for testing purpose, we can call it as many times
as we want.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>