add kicked flag to the seasons table
We will need it to tell whether the season should
be kicked when somebody makes a call. Kick should
occur only once in the season and it should be kicked
before the first transaction in the season happens,
so we need a flag to tell us if we have been already
kicked.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>